Move back code style config for backward compatibility
